requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
PyJWT
